1.用户登录; 2.fix bug;

+++ src/main/java/com/navi/controller/AppController.java
@@ -123,7 +123,7 @@ public class AppController {
      */
     @GetMapping("/getAppsByCT")
     @ResponseBody
-    public Object getAppsByCT(int ct, int curPage){
+    public Object getAppsByCT(long ct, int curPage){
         //todo:
         log.warn("enter showAppsByCT, curPage:" + curPage + ", ct:" + ct);
